+++ packages/dart_mappable/test/copy_with/copy_with_subtype_test.dart
@@ -5,24 +5,33 @@ import 'copy_with_test.dart';
 
 part 'copy_with_subtype_test.mapper.dart';
 
-@MappableClass(discriminatorKey: 'type', generateMethods: GenerateMethods.copy | GenerateMethods.stringify)
+@MappableClass(
+  discriminatorKey: 'type',
+  generateMethods: GenerateMethods.copy | GenerateMethods.stringify,
+)
 abstract class Animal with AnimalMappable {
   String name;
   Animal(this.name);
 }
 
-@MappableClass(generateMethods: GenerateMethods.copy | GenerateMethods.stringify)
+@MappableClass(
+  generateMethods: GenerateMethods.copy | GenerateMethods.stringify,
+)
 class Cat extends Animal with CatMappable {
   String color;
   Cat(super.name, this.color);
 }
 
-@MappableClass(generateMethods: GenerateMethods.copy | GenerateMethods.stringify)
+@MappableClass(
+  generateMethods: GenerateMethods.copy | GenerateMethods.stringify,
+)
 class Dog extends Animal with DogMappable {
   Dog(super.name);
 }
 
-@MappableClass(generateMethods: GenerateMethods.copy | GenerateMethods.stringify)
+@MappableClass(
+  generateMethods: GenerateMethods.copy | GenerateMethods.stringify,
+)
 class Zoo with ZooMappable {
   Animal animal;
 
@@ -43,7 +52,8 @@ class B with BMappable {
 }
 
 @MappableClass(generateMethods: GenerateMethods.copy)
-class ItemContainer2<T> extends ItemContainer<T> with ItemContainer2Mappable<T> {
+class ItemContainer2<T> extends ItemContainer<T>
+    with ItemContainer2Mappable<T> {
   ItemContainer2(super.items);
 }