+++ packages/dart_mappable/test/copy_with/copy_with_test.dart
@@ -59,7 +59,8 @@ class KeyedItemList<K, T> extends ItemList<T> with KeyedItemListMappable<K, T> {
 }
 
 @MappableClass()
-class ComparableItemList<T extends Comparable> extends ItemList<T> with ComparableItemListMappable<T> {
+class ComparableItemList<T extends Comparable> extends ItemList<T>
+    with ComparableItemListMappable<T> {
   ComparableItemList(super.items);
 }
 
@@ -93,7 +94,9 @@ void main() {
       p2 = p2.copyWith.car(brand: null);
       expect(p2.car.brand, equals(null));
 
-      Person p3 = person.copyWith.car.$update((c) => Car(c.brand, '${c.model}_xx'));
+      Person p3 = person.copyWith.car.$update(
+        (c) => Car(c.brand, '${c.model}_xx'),
+      );
       expect(p3.car.brand, equals(person.car.brand));
       expect(p3.car.model, equals('A8_xx'));
     });
@@ -103,7 +106,10 @@ void main() {
       var bmw = Brand('BMW');
       var porsche = Brand('Porsche');
       var cars = [Car(audi, 'A9'), Car(bmw, 'M4')];
-      var salesRep = {audi: Person('Max', Car(audi, 'A9')), bmw: Person('Cathy', Car(bmw, 'M3'))};
+      var salesRep = {
+        audi: Person('Max', Car(audi, 'A9')),
+        bmw: Person('Cathy', Car(bmw, 'M3')),
+      };
 
       var dealership = Dealership([...cars], {...salesRep});
 
@@ -112,16 +118,32 @@ void main() {
         equals(Dealership([Car(audi, 'A8'), Car(bmw, 'M4')], salesRep)),
       );
 
-      expect(dealership.copyWith.cars.add(Car(audi, 'A8')), equals(Dealership([...cars, Car(audi, 'A8')], salesRep)));
-
       expect(
-        dealership.copyWith.salesRep.get(audi)!(name: 'Tony'),
-        equals(Dealership(cars, {...salesRep, audi: Person('Tony', Car(audi, 'A9'))})),
+        dealership.copyWith.cars.add(Car(audi, 'A8')),
+        equals(Dealership([...cars, Car(audi, 'A8')], salesRep)),
       );
 
       expect(
-        dealership.copyWith.salesRep.put(porsche, Person('Justus', Car(porsche, '911'))),
-        equals(Dealership(cars, {...salesRep, porsche: Person('Justus', Car(porsche, '911'))})),
+        dealership.copyWith.salesRep.get(audi)!(name: 'Tony'),
+        equals(
+          Dealership(cars, {
+            ...salesRep,
+            audi: Person('Tony', Car(audi, 'A9')),
+          }),
+        ),
+      );
+
+      expect(
+        dealership.copyWith.salesRep.put(
+          porsche,
+          Person('Justus', Car(porsche, '911')),
+        ),
+        equals(
+          Dealership(cars, {
+            ...salesRep,
+            porsche: Person('Justus', Car(porsche, '911')),
+          }),
+        ),
       );
     });